Create a new JDBC driver framework, Avatica.
Pieces of Optiq's JDBC driver that do not depend on Optiq are now in the "avatica" module, in package "net.hydromatic.avatica". Avatica does not depend on anything besides the standard Java library.

Optiq's JDBC driver overrides a few virtual methods in the Avatica JDBC driver. It should now be be possible to create a remote JDBC driver for Optiq (and other databases) by creating a command-handler that runs over a protocol such as Protobuf or Thrift.