Improve algorithm to look for tiles that match a query, to allow inexact matches.
Look for materialized tiles that can be rolled up before creating a tile that is an exact match for the query. If there is a tile that has the right dimensionality but the wrong measures, in the worst case create a new tile with the union of the measures. If there are multiple tiles that can be rolled up, choose the one with the fewest rows.

Add attribute "algorithmMaxMillis" to JSON lattice element, and improve documentation for JSON model.

Make class TileKey top-level.

Add AggregateCall.create, which can infer the type of the call before creating it. This allows us to make SqlSumEmptyIsZeroAggFunction a singleton, by removing its type argument.

Deprecate interface Aggregation; new code should use SqlAggFunction.